:clown_face: Create react_library, move DateUtils to it
