+++ ark_nova_stats/frontend/src/components/DatabaseStatistics.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import Stats from '../types/Stats';
-import {getDate} from '../../../../skeleton/frontend/src/DateUtils';
+import {getDate} from '../../../../react_library/DateUtils';
 
 type DatabaseStatisticsParams = {
     stats: Stats | undefined,
