+++ react_library/globals.d.ts
@@ -1,0 +1,6 @@
+declare module "*.png" {
+    const value: any;
+    export = value;
+ }
+
+ declare module '*.css';
