+++ react_library/jest.config.json
@@ -1,0 +1,10 @@
+{
+  "bail": 1,
+  "verbose": true,
+  "testEnvironment": "jsdom",
+  "setupFilesAfterEnv": ["./setupTests.js"],
+  "moduleNameMapper": {
+    "\\.(jpg|ico|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "<rootDir>/mocks/fileMock.js",
+    "\\.(css|less)$": "<rootDir>/mocks/fileMock.js"
+  }
+}
