+++ react_library/package.json
@@ -1,0 +1,31 @@
+{
+  "name": "@monorepo/react_library",
+  "version": "0.1.0",
+  "private": true,
+  "type": "module",
+  "scripts": {
+    "tailwind:css": "bazel build //react_library:tailwindcss",
+    "start": "PORT=$REACT_APP_PORT react-scripts start",
+    "build": "bazel build //react_library/...",
+    "build-and-start": "npm run tailwind:css && npm run start",
+    "test": "bazel test //react_library/..."
+  },
+  "eslintConfig": {
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
+  }
+}
