+++ react_library/tsconfig.json
@@ -1,0 +1,21 @@
+{
+    "include": ["**/*"],
+    "exclude": ["**/*.spec.ts"],
+    "compilerOptions": {
+        "lib": ["dom", "dom.iterable", "esnext"],
+        "module": "esnext",
+        "moduleResolution": "bundler",
+        "target": "es2015",
+        "allowJs": true,
+        "allowSyntheticDefaultImports": true,
+        "esModuleInterop": true,
+        "isolatedModules": true,
+        "jsx": "react",
+        "noFallthroughCasesInSwitch": true,
+        "resolveJsonModule": true,
+        "skipLibCheck": true,
+        "strict": true,
+        "composite": true,
+        "declaration": true,
+    }
+}
